Import QEMU v1.3.1-metag3
Import QEMU v1.3.1-metag3-public-03b32b4 release from
http://imgtec.com/meta_insider/metaflow-sdk.asp.

This includes support for:
- Meta system level (softmmu) emulation
- Meta linux-user emulation
- Meta TCG backend.

Architectural features supported
- Meta 1 ATP and Meta 2 HTP cores
- DSP instruction set
- FPU instruction set
- MiniM compressed instruction set

It also includes partial hardware emulation for:
- Various Frontier Silicon Chorus 2 SoC peripherals
  - DMAC
  - IMG SPI
  - IMG PDP (Pixel Display Pipeline)
- Various Toumaz Xenif TZ1090 SoC peripherals
  - IMG MDC (DMA controller)
  - IMG SPI
  - IMG SCB (I2C)
  - IMG PDC (PowerDown Controller)
  - IMG PDP (Pixel Display Pipeline)
  - Synopsys DesignWare MMC
- METAmorph development board (01DG)
- Comet (TZ1090) METAmorph development board (01SP)
- MiniMorph development board (01TT - compatible with 01SP)
- Various external peripherals
  - Atmel Dataflash on SPI
  - SII9022A I2C HDMI encoder
  - QT5480 I2C touchscreen controller

Other features
- Direct loading of IMG LDR binaries, allowing a special version of
  meta-bootloader to be used to bypass the boot ROM and secondary loader
  so that a Kernel and command line can be loaded directly.
- Meta DA (JTAG Debug Adapter) BIOS calls
  - DA data channels (used for TTY and kernel log)
  - DA file server
  - DA LogF messages
- Trace output, including some compatibility with Meta Hardware Trace
  features.
- SAP (simulator access protocol) stub for debugging with Codescape
  Debugger.

A big thanks goes to Paul Burton for adding MiniM, FPU, and DSP support
(with associated refactoring) and various peripherals, and Paul Clothier
for implementing the Meta TCG backend.

+++ hw/kvmvapic.c
@@ -390,7 +390,7 @@ static void patch_instruction(VAPICROMState *s, CPUX86State *env, target_ulong i
     TranslationBlock *current_tb;
     target_ulong current_pc = 0;
     target_ulong current_cs_base = 0;
-    int current_flags = 0;
+    uint64_t current_flags = 0;
 
     if (smp_cpus == 1) {
         handlers = &s->rom_state.up;
